refactor: move shared methods in an utility object, update implementation of isStatecomplaintWithGoal
