+++ domain/src/main/kotlin/domain/BlockWorldDomain.kt
@@ -391,7 +391,7 @@ object BlockWorldDomain {
         val pickX = FluentBasedGoal.of(Fluents.atXArm)
         val pickXfloorY = FluentBasedGoal.of(Fluents.atXArm, Fluents.atYFloor)
         val onXY = FluentBasedGoal.of(Fluents.onXY)
-        val onXYatW = FluentBasedGoal.of(Fluents.onXY, Fluents.atWArm) // caso sfigato
+        val onXYatW = FluentBasedGoal.of(Fluents.atWArm, Fluents.onXY) // caso sfigato
         val onABC = FluentBasedGoal.of(Fluents.onCA, Fluents.onAB)
     }
